Added check if old folders should be deleted.

+++ multi_sample_factory/slurm/run.py
@@ -66,6 +66,37 @@ def parse_time_limit(time_limit: int) -> Tuple[str, str]:
     return time_str, partition
 
 
+def query_yes_no(question, default="yes"):
+    """Ask a yes/no question via raw_input() and return their answer.
+
+    "question" is a string that is presented to the user.
+    "default" is the presumed answer if the user just hits <Enter>.
+            It must be "yes" (the default), "no" or None (meaning
+            an answer is required of the user).
+
+    The "answer" return value is True for "yes" or False for "no".
+    """
+    valid = {"yes": True, "y": True, "ye": True, "no": False, "n": False}
+    if default is None:
+        prompt = " [y/n] "
+    elif default == "yes":
+        prompt = " [Y/n] "
+    elif default == "no":
+        prompt = " [y/N] "
+    else:
+        raise ValueError("invalid default answer: '%s'" % default)
+
+    while True:
+        sys.stdout.write(question + prompt)
+        choice = input().lower()
+        if default is not None and choice == "":
+            return valid[default]
+        elif choice in valid:
+            return valid[choice]
+        else:
+            sys.stdout.write("Please respond with 'yes' or 'no' " "(or 'y' or 'n').\n")
+
+
 def main():
     args = runner_argparser().parse_args(sys.argv[1:])
 
@@ -88,14 +119,16 @@ def main():
     # Remove old log and job dirs
     jobs_directory = os.path.join(args.destination, grid.name)
     logs_directory = os.path.join(args.log_dir, grid.name)
-    try:
-        shutil.rmtree(jobs_directory)
-    except OSError as e:
-        pass
-    try:
-        shutil.rmtree(logs_directory)
-    except OSError as e:
-        pass
+    if os.path.isdir(jobs_directory):
+        if query_yes_no('A jobs directory for this grid already exists. Do you want to delete it? If you answer with no, this script will abord.'):
+            shutil.rmtree(jobs_directory)
+        else:
+            return ExperimentStatus.INTERRUPTED
+    if os.path.isdir(logs_directory):
+        if query_yes_no('A log directory for this grid already exists. Do you want to delete it? If you answer with no, this script will abord.'):
+            shutil.rmtree(logs_directory)
+        else:
+            return ExperimentStatus.INTERRUPTED
     Path(jobs_directory).mkdir(parents=True, exist_ok=True)
     Path(logs_directory).mkdir(parents=True, exist_ok=True)
 
@@ -110,7 +143,8 @@ def main():
             n = 1
         for repetition in range(args.repeat):
             full_job_name = job_name + "_{0:03d}".format(repetition)
-            bash_script = grid.setup.format(partition, time_limit_str, n, full_job_name, grid.env, args.msf_dir, os.path.join(logs_directory, full_job_name) + '.log')
+            bash_script = grid.setup.format(partition, time_limit_str, n, full_job_name, grid.env, args.msf_dir,
+                                            os.path.join(logs_directory, full_job_name) + '.log')
             if grid.base_parameters != "":
                 bash_script = bash_script + " " + grid.base_parameters
             for parameter, value in zip(keys, combination):
@@ -119,7 +153,8 @@ def main():
                 else:
                     bash_script = bash_script + " --{0}={1}".format(parameter, value)
             # Add experiment and env
-            bash_script = bash_script + " --env={0} --experiment={1} --train_for_seconds={2} --train_dir={3}".format(grid.env, full_job_name, args.time_limit * 60, os.path.join(args.train_dir, grid.name))
+            bash_script = bash_script + " --env={0} --experiment={1} --train_for_seconds={2} --train_dir={3}".format(
+                grid.env, full_job_name, args.time_limit * 60, os.path.join(args.train_dir, grid.name))
 
             # Write the file
             file_path = os.path.join(jobs_directory, '{0}.sh'.format(full_job_name))
